# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Naturascan API",
        "description": "Naturascan API",
        "contact": {
            "name": "Gilles",
            "email": "gillesakakpo01@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/sortie_obstrace": {
            "get": {
                "tags": [
                    "SortieObstrace"
                ],
                "summary": "List all sortie_obstrace species",
                "operationId": "46a7a1c95ea2f63bd00b8be4817af79e",
                "responses": {
                    "200": {
                        "description": "List all sortie_obstrace species"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "SortieObstrace"
                ],
                "summary": "Create a new sortie_obstrace species",
                "operationId": "89de0a139e915d7b69159c6cf4d7e5fe",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SortieObstrace"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Create a new sortie_obstrace species"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/sortie_obstrace/{id}": {
            "get": {
                "tags": [
                    "SortieObstrace"
                ],
                "summary": "Show sortie_obstrace species by ID",
                "operationId": "a432266ccb558b0964ed6385d4897d68",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the sortie_obstrace species",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Show sortie_obstrace species by ID"
                    },
                    "404": {
                        "description": "SortieObstrace species not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "SortieObstrace"
                ],
                "summary": "Update sortie_obstrace species by ID",
                "operationId": "10279d94bae623173c4dd5b6a3b2cdab",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the sortie_obstrace species",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/SortieObstrace"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update sortie_obstrace species by ID"
                    },
                    "404": {
                        "description": "SortieObstrace species not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "SortieObstrace"
                ],
                "summary": "Delete sortie_obstrace species by ID",
                "operationId": "9eecaa647f639844f1bc36701a9da407",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the sortie_obstrace species",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Delete sortie_obstrace species by ID"
                    },
                    "404": {
                        "description": "SortieObstrace species not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "SortieObstrace": {
                "title": "SortieObstrace",
                "description": "SortieObstrace",
                "properties": {
                    "plage": {
                        "title": "plage",
                        "type": "string"
                    },
                    "nbre_observateurs": {
                        "title": "nbre_observateurs",
                        "type": "integer"
                    },
                    "suivi": {
                        "title": "suivi",
                        "type": "string"
                    },
                    "prospection_heure_debut": {
                        "title": "prospection_heure_debut",
                        "type": "string"
                    },
                    "prospection_heure_fin": {
                        "title": "prospection_heure_fin",
                        "type": "string"
                    },
                    "remark": {
                        "title": "remark",
                        "type": "string"
                    },
                    "type_bateau": {
                        "title": "type_bateau",
                        "type": "string"
                    },
                    "nom_bateau": {
                        "title": "nom_bateau",
                        "type": "string"
                    },
                    "hauteur_bateau": {
                        "title": "hauteur_bateau",
                        "type": "string"
                    },
                    "sortie_id": {
                        "title": "sortie_id",
                        "type": "integer"
                    },
                    "weather_report_id": {
                        "title": "weather_report_id",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}